Move all brokerage-specific code into `brokers` module
